Refactor PageViewModelCard --> ModelLinkCard
