+++ frontend/src/model-views/components/index.ts
@@ -3,3 +3,4 @@ export * from "./new-model";
 export * from "./page-view";
 export * from "./GeoDeepDive";
 export * from "./infinitie-scroll";
+export * from "./model-card";
