+++ frontend/src/model-views/components/model-card/index.ts
@@ -1,62 +1,13 @@
-import { Tooltip, Card, Button } from "@blueprintjs/core";
+import { Card, Button } from "@blueprintjs/core";
 import { LinkCard } from "@macrostrat/ui-components";
-import { format } from "date-fns";
-import { useModelURL } from "~/util";
 import { hyperStyled } from "@macrostrat/hyper";
 //@ts-ignore
 import styles from "./module.styl";
 
 const h = hyperStyled(styles);
 
-export const AddCard = (props) => {
-  const { onClick, model } = props;
-  return h(
-    Tooltip,
-    { content: `Select from exisitng ${model}s` },
-    h(Button, { onClick, icon: "small-plus", minimal: true })
-  );
-};
-
-export const NewModelButton = (props) => {
-  const { model } = props;
-
-  const to = useModelURL(`/${model}/new`);
-  const handleClick = (e) => {
-    e.preventDefault();
-    window.location.href = to;
-  };
-
-  const string = model.charAt(0).toUpperCase() + model.slice(1);
-
-  return h(
-    Button,
-    {
-      minimal: true,
-      intent: "success",
-      onClick: handleClick,
-    },
-    [`New ${string}`]
-  );
-};
-
-export const ModelAttributeOneLiner = (props) => {
-  const { title, content } = props;
-
-  let displayContent;
-  if (!content) {
-    displayContent = "None";
-  } else {
-    displayContent = content;
-  }
-
-  return h("span", { style: { display: "flex", alignItems: "baseline" } }, [
-    h("h4", { style: { marginRight: "4px" } }, title),
-    " ",
-    displayContent,
-  ]);
-};
-
 export const ModelLinkCard = (props) => {
+  /** A model link card for administration pages */
   const {
     minimal = true,
     className,
