+++ frontend/src/model-views/components/model-card/module.styl
@@ -1,3 +1,18 @@
+.sample-card
+  background-color: $light-gray4
+  border-radius: 5px
+  margin: 5px
+  padding: 5px 10px
+  flex-grow: 1
+  max-width: 15em
+  .location-name
+    color: $indigo3
+  .material
+    font-style italic
+  h4
+    margin-top: 0em
+    margin-bottom: 0em
+
 .indirect-link
   background-color: $light-gray5
   border-radius: 5px
