+++ frontend/src/model-views/components/page-view/module.styl
@@ -108,41 +108,6 @@
 .sample-edit-card-samhover
   background-color: yellow
 
-.indirect-link
-  background-color: $light-gray5
-  border-radius: 5px
-  border-style: dashed
-  border-color $light-gray2
-  border-width: 2px
-  margin: 5px
-  padding: 5px 10px
-  flex-grow: 1
-  max-width: 15em
-  :hover
-    background-color: $light-gray4
-  .location-name
-    color: $indigo3
-  .material
-    font-style italic
-  h4
-    margin-top: 0em
-    margin-bottom: 0em
-
-.sample-card
-  background-color: $light-gray4
-  border-radius: 5px
-  margin: 5px
-  padding: 5px 10px
-  flex-grow: 1
-  max-width: 15em
-  .location-name
-    color: $indigo3
-  .material
-    font-style italic
-  h4
-    margin-top: 0em
-    margin-bottom: 0em
-
 .project-card
   margin-bottom 1em
   color #202B33
